Revert "Add dummy REST Data"
This reverts commit 11fa997b8a62dd196bf0f0a3fed53646f8f19bef.

+++ JobTracker/src/net/greenbeansit/jobtracker/shared/EmployeeService.java
@@ -57,7 +57,7 @@ public interface EmployeeService extends DirectRestService {
 	
 	@GET
 	@Path("/{employeeId}/job/{jobId}")
-	Job getJob(@PathParam("employeeId") Long employeeId, @PathParam("jobId") Integer jobId);
+	Job getJob(@PathParam("employeeId") Long employeeId, @PathParam("jobId") Long jobId);
 
 	/**
 	 * Gets all {@link ActivityReport} for the {@link Employee} with the
